# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: VirtuProfs
basePath: /api
host: localhost:5100
schemes:
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  passport:
    type: apiKey
    in: header
    name: Authorization
  admin:
    type: apiKey
    in: header
    name: Authorization
  teacher:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  /forgot/password:
    x-swagger-router-controller: Authorization
    post:
      operationId: forgotRestore
      summary: Restore forgotten password
      tags:
        - authorization
      parameters:
        - name: data
          in: body
          description: User unique ID
          schema:
            type: object
            required:
              - token
              - pw
            properties:
              token:
                type: string
              pw:
                type: string
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
  /confirm:
    x-swagger-router-controller: Authorization
    post:
      produces:
        - application/json
      operationId: confirm
      summary: Confirm
      tags:
        - authorization
      parameters:
        - in: query
          name: token
          type: string
          description: token for confirm registration
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
...
definitions:
  Response:
    type: object
    properties:
      success:
        type: boolean
      errorCode:
        type: string
      errorStack:
        type: array
        items:
            type: object
      message:
        type: string
      data:
        type: object
    required:
      - success
...